MYR-596 Decompose the three files this change pushed past the 300-line cap
Mechanical, no behaviour change.

account_deletion_queries.go 298 -> 175: the identity transaction's statements
and the MYR-452 convergence walk move to account_deletion_identity_queries.go.
The seam already existed one layer up — account_deletion.go holds the per-step
methods, account_deletion_identity.go holds the transaction — and the queries
now follow it: statements run alone over one id outside a transaction, versus
statements run together over the whole scope inside the one that takes the row
locks.

fleetorphan/sweep.go 259 -> 273: the two DB-only passes (source C's attempt
rows, and MYR-596's tombstone purge) move to db_passes.go. They are the halves
that need no token and no network, which is why both complete when Tesla is
unreachable. Run still owns their order and still argues it at each call site.

cmd/sweep-orphan-fleet-configs/main.go 265 -> 171: the 150-line operator header
becomes doc.go, the idiomatic home for a package comment that long.
